Created unique slug generation per question
When each question is about to be saved as a new question (and by default sends out a pre-save signal), a new receiver now generates a new unique slug if the question doesn't already have one (which it will if it's being updated).

this is assigned as that instance's .slug attribute, then the question is saved as normal.