Added 'delete' and 'edit' buttons for each Question.
Created a 'QuestionActions' component to only show if author is the same as current user. Seperated to avoid bloat.

If delete is hit, a delete request is sent to that question's endpoint.

If edit is hit, the user is redirected to the 'question creator', but this time with the optional extra slug parameter, which the component now 'beforeRouteEnter' checks if there is a slug, and if so, retrieves data from that endpoint and pre-fills the textbox with the existing question's content.

edited onSubmit method to add the question's slug to the endpoint, and change the method to PUT.