Added 'like' and '404 not found' functionality.
Added a 'like' button to answers not published by the user. This can be toggled on/off with a visual indicator, and incrememnts/decrements the total likes on that answer accordingly.

Also added a catch-all path to the router index for if users enter wrong url's.

Specifically in the question component, also shown a '404' message if an attempt to retrieve a question based on the url provided was unsuccessful (wrong url).